feat(runtime): support Windows CPU placement
Discover the current Windows processor group and core topology, preserve group-aware CPU coordinates in the allocation catalog, and prefer higher-performance core classes. Pass the resolved coordinates to libkrun so every vCPU thread receives hard Windows group affinity before guest execution. Keep managed placement explicitly unsupported on macOS.

+++ crates/db/lib/entity/cpu_allocation_cpu.rs
@@ -12,7 +12,7 @@ use sea_orm::entity::prelude::*;
 pub struct Model {
     /// Host logical processor identifier and global conflict key.
     #[sea_orm(primary_key, auto_increment = false)]
-    pub logical_cpu: i32,
+    pub logical_cpu: i64,
     /// Owning allocation identifier.
     pub allocation_id: String,
     /// Possible guest vCPU index, or `None` for policy-only sibling reservations.
